openxr: new module with a RenderingMode targeting OpenXR devices
OpenXR is a standard to provides access to Augmented Reality (AR)
and Virtual Reality (VR) devices. Application should connect to
an OpenXR loader that will scan the OpenXR runtimes available on the
the host. Each runtime will communicate with its physical connected
XR device.

Be able to move freely in a 3DGS scene through a VR headset would be very
cool.

'openxr' module handles the boiler-blate to:
* connect to a Head-Mounted Display form factor XR device
* listen to session state change
* render a SIBR view into two FBO textures (one for each eye) that
  will be sent out to the connected XR device (through SwapchainImage mechanism)
* display both textures into UI window
* synchronize headset rendering framerate with the SIBR view
* handle two VR experience (free world standing, seated)

The module works on Windows and Linux over XLIB (wayland is not supported). It
has been tested with Meta Quests headsets on Windows and HTV Vive Pro
on Linux (through SteamVR OpenXR runtime).

Signed-off-by: Cedric CHEDALEUX <cedric.chedaleux@orange.com>